# docs/swagger.yaml
definitions:
...
  albums.AlbumDetailResponse:
    properties:
      created_at:
        type: integer
      description:
        type: string
      id:
        type: integer
      image_count:
        type: integer
      images:
        items:
          $ref: '#/definitions/albums.AlbumImageDTO'
        type: array
      name:
        type: string
      updated_at:
        type: integer
    type: object
...
  albums.CreateAlbumResponse:
    properties:
      created_at:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: integer
    type: object
  albums.ListAlbumsResponse:
    properties:
      albums:
        items:
          $ref: '#/definitions/albums.AlbumDTO'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  albums.createAlbumRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 100
        type: string
    required:
    - name
    type: object
  api.loginResponse:
    properties:
      access_token:
        type: string
      access_token_expiry:
        type: integer
    type: object
  api.userAuthRequestBody:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: string
    type: object
...
  images.SetRandomSourceAlbumRequest:
    properties:
      album_id:
        type: integer
      include_all_public:
        type: boolean
    type: object
  images.UpdateVisibilityRequest:
    properties:
      is_public:
        type: boolean
    required:
    - is_public
    type: object
...
  models.SystemConfigStoreRequest:
    properties:
      category:
        $ref: '#/definitions/models.ConfigCategory'
      config:
        additionalProperties: true
        type: object
      description:
        type: string
      is_default:
        type: boolean
      is_enabled:
        type: boolean
      name:
        type: string
      priority:
        type: integer
    required:
    - category
    - config
    - name
    type: object
...
info:
  contact: {}
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with username and password, returns access token
        and sets refresh token cookie
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.userAuthRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.loginResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      summary: User login
      tags:
      - auth
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user by invalidating refresh token (requires refresh_token
        and device_id cookies)
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Refresh token not found / invalid session
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      summary: User logout
      tags:
      - auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh_token and device_id cookies
      produces:
      - application/json
      responses:
        "200":
          description: Refresh token successful
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.loginResponse'
              type: object
        "401":
          description: Refresh token or device ID not found / invalid
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/admin/configs:
    get:
      consumes:
      - application/json
      description: Get list of system configurations including storage and image processing
        settings
      parameters:
      - description: 'Filter by category: storage, image_processing'
        in: query
        name: category
        type: string
      - description: Only show enabled configs
        in: query
        name: enabled_only
        type: boolean
      - description: 'Mask sensitive information (default: true)'
        in: query
        name: mask_sensitive
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Configuration list
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: List system configurations
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a new system configuration (storage or image processing)
      parameters:
      - description: Configuration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SystemConfigStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Configuration created successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request or configuration test failed
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Create configuration
      tags:
      - admin
...
  /api/v1/admin/random-source-album:
    get:
      consumes:
      - application/json
      description: Get the currently configured random image source album ID and whether
        to include all public images
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Get random source album configuration
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Configure the source album for random image selection. Use album_id=0
        for all public images, or specify an album ID
      parameters:
      - description: Random source configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/images.SetRandomSourceAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Configuration updated successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Set random source album configuration
      tags:
      - admin
  /api/v1/admin/storage/providers:
    get:
      consumes:
      - application/json
      description: Get list of all loaded storage providers
      produces:
      - application/json
      responses:
        "200":
          description: Storage provider list
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: List storage providers
      tags:
      - admin
  /api/v1/admin/storage/reload/{id}:
    post:
      consumes:
      - application/json
      description: Hot reload a storage configuration (not supported in simplified
        mode)
      parameters:
      - description: Storage config ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reload status
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Reload storage configuration
      tags:
      - admin
  /api/v1/albums:
    get:
      consumes:
      - application/json
      description: Get paginated list of user's albums with image counts and cover
        URLs
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Album list
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/albums.ListAlbumsResponse'
              type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: List albums
      tags:
      - albums
    post:
      consumes:
      - application/json
      description: Create a new album with name and optional description
      parameters:
      - description: Album creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/albums.createAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Album created successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/albums.CreateAlbumResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Create album
      tags:
      - albums
  /api/v1/albums/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an album by ID (images in the album will not be deleted)
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Album deleted successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid album ID format
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Album not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete album
      tags:
      - albums
    get:
      consumes:
      - application/json
      description: Get detailed information about an album including all images
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Album details
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/albums.AlbumDetailResponse'
              type: object
        "400":
          description: Invalid album ID format
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Album not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Get album detail
      tags:
      - albums
    put:
      consumes:
      - application/json
      description: Update album name and description
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Album update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/albums.UpdateAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Album updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/albums.UpdateAlbumResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Album not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Update album
      tags:
      - albums
  /api/v1/albums/{id}/images:
    post:
      consumes:
      - application/json
      description: Add multiple images to an existing album
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image IDs to add
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/albums.AddImagesToAlbumRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Images added to album successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Album or images not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Add images to album
      tags:
      - albums
  /api/v1/albums/{id}/images/{imageId}:
    delete:
      consumes:
      - application/json
      description: Remove a single image from an album (image itself is not deleted)
      parameters:
      - description: Album ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image ID
        in: path
        name: imageId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Image removed from album successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Album or image not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove image from album
      tags:
      - albums
  /api/v1/dashboard/stats:
    get:
      consumes:
      - application/json
      description: Get dashboard statistics including image count, storage usage,
        and recent activity
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard statistics
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Get dashboard statistics
      tags:
      - dashboard
  /api/v1/dashboard/stats/refresh:
    post:
      consumes:
      - application/json
      description: Force refresh the dashboard statistics cache
      produces:
      - application/json
      responses:
        "200":
          description: Stats refreshed successfully
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Refresh dashboard statistics
      tags:
      - dashboard
  /api/v1/images:
    post:
      consumes:
      - application/json
      description: Get paginated list of images with optional filters
      parameters:
      - description: Filter and pagination parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/images.ImageRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: Image list
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/images.ImageListResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: List images
      tags:
      - images
  /api/v1/images/{identifier}:
    delete:
      consumes:
      - application/json
      description: Delete a single image by its identifier
      parameters:
      - description: Image identifier
        in: path
        name: identifier
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Image deleted successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid identifier
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Image not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete single image
      tags:
      - images
  /api/v1/images/{identifier}/visibility:
    put:
      consumes:
      - application/json
      description: Toggle image visibility between public and private
      parameters:
      - description: Image identifier
        in: path
        name: identifier
        required: true
        type: string
      - description: Visibility settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/images.UpdateVisibilityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Image visibility updated successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Image not found
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Update image visibility
      tags:
      - images
  /api/v1/images/delete:
    post:
      consumes:
      - application/json
      description: Delete multiple images by their identifiers in a single request
      parameters:
      - description: List of image identifiers to delete
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/images.DeleteRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: Delete request processed successfully
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete multiple images
      tags:
      - images
  /api/v1/images/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a single image file with optional storage strategy and visibility
        settings
      parameters:
      - description: Image file to upload
        in: formData
        name: file
        required: true
        type: file
      - description: Storage strategy ID
        in: formData
        name: strategy_id
        type: string
      - description: 'Whether the image is public (default: true)'
        in: formData
        name: is_public
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Upload successful
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Invalid form data or too many files
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "413":
          description: File too large
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - ApiKeyAuth: []
      summary: Upload single image
      tags:
      - images
...
swagger: "2.0"